+++ R/register_umap.R
@@ -1,6 +1,6 @@
 #######################################################################
 # seriation - Infrastructure for seriation
-# Copyright (C) 2011 Michael Hahsler, Christian Buchta and Kurt Hornik
+# Copyright (C) 2015 Michael Hahsler, Christian Buchta and Kurt Hornik
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -16,16 +16,29 @@
 # with this program; if not, write to the Free Software Foundation, Inc.,
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
+## register seriation based on umap
 
-ggVAT <- function(x, upper_tri = TRUE, lower_tri = TRUE, ...) {
-  if (!inherits(x, "dist"))
-    stop("x needs to be of class 'dist'!")
-  ggpimage(x, seriate(x, "VAT"), upper_tri = upper_tri, lower_tri = lower_tri, ...)
-}
+register_umap <- function() {
+  check_installed("umap")
 
-ggiVAT <- function(x, upper_tri = TRUE, lower_tri = TRUE, ...) {
-  if (!inherits(x, "dist"))
-    stop("x needs to be of class 'dist'!")
-  x <- path_dist(x)
-  ggpimage(x, seriate(x, "VAT"), upper_tri = upper_tri, lower_tri = lower_tri, ...)
+  .contr <- unclass(umap::umap.defaults)
+  .contr$n_components <- 1
+  .contr$input <- "dist"
+
+  umap_order <- function(x, control) {
+    control <- .get_parameters(control, .contr)
+
+    class(control) <- class(umap::umap.defaults)
+
+    embedding <- umap::umap(as.matrix(x), config = control)
+    order(embedding$layout)
+  }
+
+  set_seriation_method(
+    "dist",
+    "umap",
+    umap_order,
+    "Use 1D Uniform manifold approximation and projection (UMAP) embedding to create an order",
+    .contr
+  )
 }